+++ packages/core/ios/RNGoogleMobileAds/RNGoogleMobileAdsConsentModule.mm
@@ -23,7 +23,7 @@
 #include <UserMessagingPlatform/UserMessagingPlatform.h>
 #endif
 #ifdef RCT_NEW_ARCH_ENABLED
-#import "RNGoogleMobileAdsSpec.h"
+#import <RNGoogleMobileAdsSpec/RNGoogleMobileAdsSpec.h>
 #endif
 #import "RNGoogleMobileAdsConsentModule.h"
 #import "common/RNSharedUtils.h"