+++ packages/core/ios/RNGoogleMobileAds/RNGoogleMobileAdsModule.mm
@@ -22,7 +22,7 @@
 
 #import "RNGoogleMobileAdsModule.h"
 #ifdef RCT_NEW_ARCH_ENABLED
-#import "RNGoogleMobileAdsSpec.h"
+#import <RNGoogleMobileAdsSpec/RNGoogleMobileAdsSpec.h>
 #endif
 #import "common/RNSharedUtils.h"